Fix printk types in SPI flash drivers
- use %zu instead of %zd for size_t (%zd is for ssize_t)
- use %x instead of %lx for u32
- break some long lines to avoid commit hook trouble

Reviewed-on: http://review.coreboot.org/1041
Tested-by: build bot (Jenkins)

+++ src/drivers/spi/sst.c
@@ -195,7 +195,7 @@ sst_write(struct spi_flash *flash, u32 offset, size_t len, const void *buf)
 
 done:
 	printk(BIOS_INFO, "SF: SST: program %s %zu bytes @ 0x%lx\n",
-	      ret ? "failure" : "success", len, offset - actual);
+	      ret ? "failure" : "success", len, (unsigned long)offset - actual);
 
 	spi_release_bus(flash->spi);
 	return ret;
